fix: use good practices from testing lib
Async isn't necessary for the first test suit of hello.test.

Regards to the second one: From some reason this test breaks if not using async/await, even though it shouldn't be used with userEvent considering it doesn't return a promise.

Here we are also using screen for the second suite to query the elements as encouraged by testing lib, and not destructuring.