fix: prevent tests from stalling if an error occurs (fix #374)
The serialization logic of error objects would mutate the error object in-place,
retaining its original constructor and leaving non-enumerable properties intact.
When an object of type `Error` is sent over a `MessagePort` an issue may occur
which resulted in the tests stalling. The exact reason as to why the tests get
stuck before this fix are unclear, but serializing the `Error` instance into
a plain object avoids the issue.

This commit changes the serialization logic to create a new object without
its original constructor and then copies over all properties to the clone, which
prevents the tests from stalling. An integration fixture is added in the `fails`
workspace to prevent regressions.

These changes make the custom `nameStr` and `stackStr` properties redundant,
as `name` and `stack` are now copied over when traversing the prototype chain of
an `Error` object. These properties are not removed as that would be cleanup work
outside the scope of this fix.

Fixes #374