fix: auto-mock based on descriptor list, not property keys (fix #1671)
These two methods seem like they should be the same, but they're not.
Some objects, especially native Node exports, behave strangely; they
have properties that are enumerated by `Object.getOwnPropertyNames()`
and/or `Object.getOwnPropertySymbols()`, but when you try to get the
descriptor for that property using `Object.getOwnPropertyDescriptor()`
it returns `undefined`. To combat this, instead of getting the property
names / symbols manually, we instead offload that work to the JS engine
itself via `Object.getOwnPropertyDescriptors`, and just iterate through
the result.

Fixes #1671