docs: correct jsdom typo (#2011)
I tested `jsdon` to confirm it does not work. When I tried it, I was prompted to install undefined. It does work when the value is `jsdom` so I made that change here.

Perhaps there should be some error handling for when an unknown environment is specified rather than letting it attempt to install undefined?