docs: update Web Test Runner comparison (#2349)
Update the comparison to Web Test Runner comparison to a new and maintained package. This will give a more honest comparison and allows developers to make more informed decision in which test framework fits them best. Also note that in the new plugin, features such as changing the viewport do work.