Make it more explicit where options go
It wasn't clear to me where the options go and I think this change would help clarify. I think the main reason for the confusion is that

```
test: {
  exclude: [..]
}
```

Reads as "test the files in this section", not "specify your testing options here". The term "options" is also used below, but not above making it unclear where the options go. Hopefully this will help others in the same situation as me.