fix(vitest): test deep dependencies change detection
`--changed` flag is not correctly looking up deep dependencies, as such
the changed flag does not trigger specs if a sub dependency is changed.

Alters the logic to add the collected dependency after recursive `addImports`
is called, otherwise the logic returns early because that dependency
has already been added to the collected deps set. So it will not
look deeply.