fix: Default to "run" (not "watch") when stdin is not a TTY
As originally proposed here: https://github.com/vitest-dev/vitest/discussions/7672

If stdin is not a TTY, we are probably running in a non-interactive context, and thus "watch" mode is likely to hang forever as the user is not around to press 'q' to quit.

This can come up in particular when using tools or AI agents that want to run the tests using `npm run test`, but where the "test" script has been define as simply `vitest`. Of course, an alternative solution would be to change the script to `vitest run`, but this makes the experience worse for humans running `npm run test`.

I was unable to figure out how to test this change. In the PR review, it was suggested that `isTTY` could be overridden in a test fixture, as done in `test/watch/fixtures/single-failed/vitest.config.ts`. This does not actually appear to work:
- The `vite.config.ts` file appears to run in a different process than the CLI under test, so the assignment has no effect.
- Indeed, when I tried removing the lines from the `single-failed` fixture, or even overriding isTTY to false instead, this did not break the test `reporter-failed.test.ts` which uses this fixture. So it appears this existing fixture is not a good example as it is not actually doing what it looks like.