test: Add unit tests for headless config value inheritance (Playwright provider only)
Adds tests for vitest-dev/vitest#7661, but I suspect that the issue is not limited to only the Playwright provider – the workspace resolution operates at a higher level to generate child projects for each browser instance.

todo(delilahwu): Add more tests: webdriver perhaps?
todo(delilahwu): Need to spy on the headless value passed to the playwright execution, or is that out of scope?