fix(vite-node): unable to handle errors where sourcemap mapping empty
https://github.com/vitejs/vite/pull/19226 introduced a change where by empty source
mappings are not combined. As such when vitest has an internal error caused
such as using jsdom feature that is not implemented which logs an error
it tries to map the source map when the mapping is empty which results
in an error such as:

```
Cannot read properties of undefined (reading 'map')
 ❯ VirtualConsole.<anonymous> node_modules/.pnpm/jsdom@26.1.0/node_modules/jsdom/lib/jsdom/virtual-console.js:29:53
 ❯ module.exports node_modules/.pnpm/jsdom@26.1.0/node_modules/jsdom/lib/jsdom/browser/not-implemented.js:12:26
 ❯ HTMLAudioElementImpl.load node_modules/.pnpm/jsdom@26.1.0/node_modules/jsdom/lib/jsdom/living/nodes/HTMLMediaElement-impl.js:112:5
 ❯ HTMLAudioElement.load node_modules/.pnpm/jsdom@26.1.0/node_modules/jsdom/lib/jsdom/living/generated/HTMLMediaElement.js:109:34
```

When the error originates from a file without source mapping such as `x.vue`
it errors instead of logging original error.