fix(ui): allow relative paths for coverage URL in Vitest UI
This fix removes the leading forward slash (`/`) on the computed `coverageUrl` value for the Vitest UI coverage html report preview.
Before this fix, if the built html coverage report files are deployed to a subdirectory (e.g. `www.vitest.dev/some/sub/directory/`), the coverageURL will not point to the correct location of the coverage report. This fix allows for the files to be referenced relatively to where the main Vitest UI html files are deployed, regardless of whether they are deployed at the root or in a sub route of a server.