refactor: renaming
PoolRuntime => PoolRunner (the runtime is the what is running, a bit confusing to mix them up. It is kind of a worker, but we provide a worker to it, so it doesn't make sense to mix them up, it's more of an orchestrator/runner on top of the worker)
PoolRuntimeWorker => PoolWorker (we always had a runtime worker, but we also always had "maxWorkers" that referred to the previously named PoolRuntime, basically. I think the new name better describes its function).
PoolRuntimeOptions => PoolOptions (we pass them down to both the runner and the worker, so let's have a shared name then)