fix(browser): locator which does not locate any element is not visible
When using an assertion such as `expect(page.getByTestId('not-existing')).not.toBeVisible()`, the assertion used to fail because the assertion tried to get the element of the locator, which threw an exception.
After this fix, the assertion will pass, because an element that does not exist (and which is thus not in the document) is de facto not visible.
It allows the developer to use `not.toBeVisible()` without having to care if the element is hidden in the DOM or removed from the DOM.